+++ site/.vitepress/config.ts
@@ -105,6 +105,7 @@ export default defineConfig({
 			{
 				text: "Project",
 				items: [
+					{ text: "What's new in 0.5.1", link: "/whats-new-0.5.1" },
 					{ text: "What's new in 0.5.0", link: "/whats-new-0.5.0" },
 					{ text: "What's new in 0.4.3", link: "/whats-new-0.4.3" },
 					{ text: "What's new in 0.4", link: "/whats-new-0.4.0" },